update instructions after re-organization
