+++ src/murthy_lab_to_nwb/cowley2022mapping/__init__.py
@@ -1,2 +1,1 @@
-from .cowley2022mapping_courtship_nwbconverter import Cowley2022MappingCourtshipNWBConverter
-from .cowley2022mapping_imaging_nwbconverter import Cowley2022MappingImagingNWBConverter
+from .cowley2022mapping_nwbconverter import Cowley2022MappingCourtshipNWBConverter, Cowley2022MappingImagingNWBConverter
