+++ src/murthy_lab_to_nwb/li2022ecephys/li2022ecephysinterface.py
@@ -38,7 +38,6 @@ class Li2022EcephysRecordingSegment(BaseRecordingSegment):
 
         return traces
 
-
 class Li2022EcephysRecording(BaseRecording):
     """
